requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.2.10
boto3==1.16.4
botocore==1.19.4
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
comtele-sdk==1.1.5
cryptography==3.1.1
cssselect2==0.4.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-base64field==1.0
django-crispy-forms==1.9.2
django-extra-fields==3.0.2
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-localflavor==3.0.1
django-object-actions==3.0.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-storages==1.10.1
djangorestframework==3.12.1
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jmespath==0.10.0
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-stdnum==1.15
pytz==2020.1
reportlab==3.5.56
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.25.11
virtualenv==20.3.1
WeasyPrint==52.1
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
gunicorn==20.0.4
